@@ -46,6 +46,7 @@ impl MirrorCmd {
             self.low,
             self.high,
             self.allow_mirror_gaps,
+            &args.storage_config,
         )
         .await?;
 
@@ -53,8 +54,7 @@ impl MirrorCmd {
             source: self.src.clone(),
             concurrency: args.concurrency,
             skip_optional: args.skip_optional,
-            max_retries: args.max_retries,
-            initial_backoff_ms: args.initial_backoff_ms,
+            storage_config: args.storage_config,
         };
 
         let pipeline = Arc::new(
