@@ -1,19 +1,26 @@
 //! Pipeline for processing Stellar History Archives
 //!
 //! The pipeline coordinates parallel processing of archive checkpoints and their files.
+//! Retry, concurrency limiting, timeout handling, and bandwidth throttling are delegated
+//! to the storage layer (`OpenDAL`).
 
 use crate::{
     history_format::{self, bucket_path, checkpoint_path, HistoryFileState},
-    storage::ErrorClass,
+    storage::StorageConfig,
+    utils::RetryState,
 };
+use bytes::Buf;
 use futures_util::{
     future::{join, join3, join_all},
     stream, StreamExt,
 };
 use lru::LruCache;
-use std::sync::{Arc, Mutex};
+use opendal::{Buffer, Reader};
+use std::sync::{
+    atomic::{AtomicUsize, Ordering},
+    Arc, Mutex,
+};
 use thiserror::Error;
-use tokio::{io::AsyncReadExt, sync::Semaphore};
 use tracing::{debug, error, info};
 
 /// Pipeline errors
@@ -45,35 +52,31 @@ const PROGRESS_REPORTING_FREQUENCY: usize = 100;
 #[async_trait::async_trait]
 pub trait Operation: Send + Sync + 'static {
     /// Get the checkpoint bounds for this operation
-    /// Returns (lower_bound, upper_bound) checkpoints to process
+    /// Returns (`lower_bound`, `upper_bound`) checkpoints to process
     async fn get_checkpoint_bounds(
         &self,
         source: &crate::storage::StorageRef,
-        max_retries: u32,
-        initial_backoff_ms: u64,
     ) -> Result<(u32, u32), Error>;
 
     /// Process an object by streaming its content from the provided reader.
-    /// Only called when existence_check_only() returns false.
+    /// Only called when `existence_check_only()` returns false.
     ///
     /// Returns:
     /// - Ok(()) on success
-    /// - Err(e) where e.class == Retry for transient errors
-    /// - Err(e) where e.class == Fatal/NotFound for permanent errors
-    async fn process_object(
-        &self,
-        path: &str,
-        reader: crate::storage::BoxedAsyncRead,
-    ) -> Result<(), crate::storage::Error>;
+    /// - Err(e) on failure (storage layer handles retries internally)
+    async fn process_object(&self, path: &str, reader: Reader)
+        -> Result<(), crate::storage::Error>;
+
+    /// Process an object from an in-memory buffer.
+    /// Used when the content is already loaded (e.g., history files that are parsed).
+    /// Each operation handles success/failure recording internally.
+    async fn process_buffer(&self, path: &str, buffer: Buffer);
 
     /// Record a successful file processing
     fn record_success(&self, path: &str);
 
     /// Record a failed file processing
     async fn record_failure(&self, path: &str);
-
-    /// Record a retry attempt
-    fn record_retry(&self);
 
     /// Record a skipped file (already exists at destination)
     fn record_skipped(&self, path: &str);
@@ -104,26 +107,12 @@ pub trait Operation: Send + Sync + 'static {
 pub struct PipelineConfig {
     /// Source archive URL, as provided by the user
     pub source: String,
-    /// Number of concurrent workers for processing
+    /// Number of concurrent checkpoint workers
     pub concurrency: usize,
     /// Whether to skip optional SCP files
     pub skip_optional: bool,
-    /// Maximum number of HTTP retry attempts
-    pub max_retries: u32,
-    /// Initial backoff in milliseconds for HTTP retries
-    pub initial_backoff_ms: u64,
-}
-
-impl Default for PipelineConfig {
-    fn default() -> Self {
-        Self {
-            source: String::new(),
-            concurrency: 64,
-            skip_optional: false,
-            max_retries: 3,
-            initial_backoff_ms: 100,
-        }
-    }
+    /// Storage layer configuration (retry, timeout, bandwidth, etc.)
+    pub storage_config: StorageConfig,
 }
 
 pub struct Pipeline<Op: Operation> {
@@ -131,72 +120,32 @@ pub struct Pipeline<Op: Operation> {
     config: PipelineConfig,
     src_store: crate::storage::StorageRef,
     bucket_lru: Mutex<LruCache<String, ()>>,
-    io_permits: Semaphore,
+    progress_tracker: AtomicUsize,
 }
 
-use crate::utils::RetryState;
-
 impl<Op: Operation> Pipeline<Op> {
-    /// Handle a storage error with retry logic.
-    /// Returns `false` if we should give up, `true` if we should retry.
-    async fn maybe_backoff_for_retry(
-        &self,
-        path: &str,
-        error: &crate::storage::Error,
-        retry: &mut RetryState,
-        action: &str,
-    ) -> bool {
-        match error.class {
-            ErrorClass::Retry => {
-                self.operation.record_retry();
-                if !retry.record_attempt() {
-                    error!(
-                        "Failed to {} {} after {} attempts: {}",
-                        action, path, retry.attempt, error
-                    );
-                    self.operation.record_failure(path).await;
-                    return false; // give up
-                }
-                debug!(
-                    "Retrying {} (attempt {}/{}): {}, backing off {}ms",
-                    path, retry.attempt, retry.max_retries, error, retry.backoff_ms
-                );
-                retry.backoff().await;
-                true // retry
-            }
-            ErrorClass::Fatal | ErrorClass::NotFound => {
-                error!("Failed to {} {}: {}", action, path, error);
-                self.operation.record_failure(path).await;
-                false // give up
-            }
-        }
-    }
-
     /// Create a new pipeline
     pub async fn new(operation: Op, config: PipelineConfig) -> Result<Self, Error> {
-        let src_store = crate::storage::from_url(&config.source)
-            .await
-            .map_err(|e| {
-                std::io::Error::new(
-                    std::io::ErrorKind::Other,
-                    format!(
+        let src_store =
+            crate::storage::from_url_with_config(&config.source, &config.storage_config)
+                .await
+                .map_err(|e| {
+                    std::io::Error::other(format!(
                         "Failed to create storage backend for {}: {}",
                         config.source, e
-                    ),
-                )
-            })?;
+                    ))
+                })?;
 
         let bucket_lru = Mutex::new(LruCache::new(
             std::num::NonZeroUsize::new(BUCKET_LRU_CACHE_SIZE).unwrap(),
         ));
-        let io_permits = Semaphore::new(config.concurrency);
 
         Ok(Self {
             operation,
             config,
             src_store,
             bucket_lru,
-            io_permits,
+            progress_tracker: AtomicUsize::new(0),
         })
     }
 
@@ -204,11 +153,7 @@ impl<Op: Operation> Pipeline<Op> {
         // Get checkpoint bounds from the operation
         let (lower_bound, upper_bound) = self
             .operation
-            .get_checkpoint_bounds(
-                &self.src_store,
-                self.config.max_retries,
-                self.config.initial_backoff_ms,
-            )
+            .get_checkpoint_bounds(&self.src_store)
             .await?;
 
         let total_count = history_format::count_checkpoints_in_range(lower_bound, upper_bound);
@@ -229,11 +174,16 @@ impl<Op: Operation> Pipeline<Op> {
 
         // Process checkpoints concurrently, limiting to config.concurrency at a time
         stream::iter(checkpoints)
-            .enumerate()
-            .for_each_concurrent(self.config.concurrency, |(i, fut)| async move {
+            .for_each_concurrent(self.config.concurrency, |fut| async {
                 fut.await;
-                if i % PROGRESS_REPORTING_FREQUENCY == 0 || i == total_count - 1 {
-                    info!("Progress: {}/{} checkpoints processed", i + 1, total_count);
+                let completed = self.progress_tracker.fetch_add(1, Ordering::Relaxed) + 1;
+                if completed.is_multiple_of(PROGRESS_REPORTING_FREQUENCY)
+                    || completed == total_count
+                {
+                    info!(
+                        "Progress: {}/{} checkpoints processed",
+                        completed, total_count
+                    );
                 }
             })
             .await;
@@ -262,79 +212,71 @@ impl<Op: Operation> Pipeline<Op> {
         }
     }
 
+    /// Process a history file for a checkpoint.
+    /// Downloads and parses the history JSON, then processes buckets.
+    /// Retries are handled at this level to avoid file corruption from partial streams.
     async fn process_history(self: Arc<Self>, checkpoint: u32) {
-        let history_path = checkpoint_path("history", checkpoint);
-        let mut retry = RetryState::new(self.config.max_retries, self.config.initial_backoff_ms);
+        use futures_util::TryStreamExt;
 
-        // Download the history file
+        let history_path = checkpoint_path("history", checkpoint);
+
+        // Retry state for downloading the history file
+        let mut retry_state = RetryState::new(
+            self.config.storage_config.max_retries as u32,
+            self.config.storage_config.retry_min_delay.as_millis() as u64,
+        );
+
+        // Download the history file with retries at this level
+        // Use into_stream to handle chunked transfer encoding properly
         let buffer = loop {
-            let download_result: Result<Vec<u8>, crate::storage::Error> = async {
-                let mut reader = self.src_store.open_reader(&history_path).await?;
-                let mut buffer = Vec::new();
-                reader
-                    .read_to_end(&mut buffer)
+            let download_result: Result<Buffer, crate::storage::Error> = async {
+                let reader = self.src_store.open_reader(&history_path).await?;
+                // Convert to stream and collect all chunks into a single buffer
+                let stream = reader
+                    .into_stream(..)
                     .await
-                    .map_err(|e| crate::storage::Error::retry(format!("Read error: {}", e)))?;
+                    .map_err(|e| crate::storage::from_opendal_error(e, "Stream error"))?;
+                let chunks: Vec<Buffer> = stream
+                    .try_collect()
+                    .await
+                    .map_err(|e| crate::storage::from_opendal_error(e, "Read error"))?;
+                // Merge chunks into a single buffer
+                let buffer: Buffer = chunks.into_iter().flatten().collect();
                 Ok(buffer)
             }
             .await;
 
             match download_result {
-                Ok(bytes) => break bytes,
+                Ok(buf) => break buf,
                 Err(e) => {
-                    if !self
-                        .maybe_backoff_for_retry(&history_path, &e, &mut retry, "download")
-                        .await
-                    {
-                        return;
+                    if retry_state.should_retry(&e, "download history file", &history_path) {
+                        retry_state.backoff().await;
+                        continue;
                     }
+                    self.operation.record_failure(&history_path).await;
+                    return;
                 }
             }
         };
 
         // Parse and validate
         let parse_result: Result<HistoryFileState, Error> = (|| {
-            let state: HistoryFileState = serde_json::from_slice(&buffer).map_err(|e| {
-                history_format::Error::InvalidJson {
+            let state: HistoryFileState = serde_json::from_reader(buffer.clone().reader())
+                .map_err(|e| history_format::Error::InvalidJson {
                     path: history_path.clone(),
                     error: e.to_string(),
-                }
-            })?;
+                })?;
             state.validate()?;
             Ok(state)
         })();
 
         match parse_result {
             Ok(state) => {
-                // For existence-only checks, we already downloaded and parsed the history file,
-                // so just record success. For mirror, we need to write it to destination.
-                if self.operation.existence_check_only() {
-                    self.operation.record_success(&history_path);
-                    self.clone().process_buckets(state).await;
-                } else {
-                    // Write the history file (we already have content in buffer) and process buckets concurrently
-                    let history_path_clone = history_path.clone();
-                    let write_history = async {
-                        let reader = Box::new(std::io::Cursor::new(buffer))
-                            as crate::storage::BoxedAsyncRead;
-                        match self
-                            .operation
-                            .process_object(&history_path_clone, reader)
-                            .await
-                        {
-                            Ok(()) => self.operation.record_success(&history_path_clone),
-                            Err(e) => {
-                                error!(
-                                    "Failed to write history file {}: {}",
-                                    history_path_clone, e
-                                );
-                                self.operation.record_failure(&history_path_clone).await;
-                            }
-                        }
-                    };
-                    let buckets = self.clone().process_buckets(state);
-                    let _ = join(write_history, buckets).await;
-                }
+                // Process the history file and process buckets concurrently
+                let history_path_clone = history_path.clone();
+                let process_history = self.operation.process_buffer(&history_path_clone, buffer);
+                let buckets = self.clone().process_buckets(state);
+                let _ = join(process_history, buckets).await;
             }
             Err(e) => {
                 error!(
@@ -373,31 +315,33 @@ impl<Op: Operation> Pipeline<Op> {
         join_all(bucket_futures).await;
     }
 
+    /// Process a single file (bucket, ledger, transactions, results, scp).
+    /// All retries are handled at this level - no automatic retries in `OpenDAL`.
     async fn process_file(self: Arc<Self>, path: String) {
-        let mut retry = RetryState::new(self.config.max_retries, self.config.initial_backoff_ms);
-
-        // Acquire I/O permit
-        let _permit = self.io_permits.acquire().await.unwrap();
-
-        loop {
-            // Pre-check: allow operation to skip without querying source
-            if let Some(result) = self.operation.pre_check(&path).await {
-                match result {
-                    Ok(()) => {
-                        debug!("Skipping: {}", path);
-                        self.operation.record_skipped(&path);
-                        return;
-                    }
-                    Err(e) => {
-                        error!("Pre-check failed for {}: {}", path, e);
-                        self.operation.record_failure(&path).await;
-                        return;
-                    }
+        // Pre-check: allow operation to skip without querying source
+        if let Some(result) = self.operation.pre_check(&path).await {
+            match result {
+                Ok(()) => {
+                    debug!("Skipping: {}", path);
+                    self.operation.record_skipped(&path);
+                    return;
+                }
+                Err(e) => {
+                    error!("Pre-check failed for {}: {}", path, e);
+                    self.operation.record_failure(&path).await;
+                    return;
                 }
             }
+        }
 
-            // For existence-only checks, we don't need to open a reader or call process_object
-            if self.operation.existence_check_only() {
+        let mut retry_state = RetryState::new(
+            self.config.storage_config.max_retries as u32,
+            self.config.storage_config.retry_min_delay.as_millis() as u64,
+        );
+
+        // For existence-only checks, we don't need to open a reader or call process_object
+        if self.operation.existence_check_only() {
+            loop {
                 match self.src_store.exists(&path).await {
                     Ok(true) => {
                         debug!("Exists: {}", path);
@@ -410,44 +354,44 @@ impl<Op: Operation> Pipeline<Op> {
                         return;
                     }
                     Err(e) => {
-                        if !self
-                            .maybe_backoff_for_retry(&path, &e, &mut retry, "check")
-                            .await
-                        {
-                            return;
+                        if retry_state.should_retry(&e, "check existence of", &path) {
+                            retry_state.backoff().await;
+                            continue;
                         }
-                    }
-                }
-            } else {
-                // Normal path - open a reader for actual content streaming
-                let reader = match self.src_store.open_reader(&path).await {
-                    Ok(reader) => reader,
-                    Err(e) => {
-                        if !self
-                            .maybe_backoff_for_retry(&path, &e, &mut retry, "open")
-                            .await
-                        {
-                            return;
-                        }
-                        continue;
-                    }
-                };
-
-                // Process the file
-                match self.operation.process_object(&path, reader).await {
-                    Ok(()) => {
-                        debug!("Processed: {}", path);
-                        self.operation.record_success(&path);
+                        self.operation.record_failure(&path).await;
                         return;
                     }
-                    Err(e) => {
-                        if !self
-                            .maybe_backoff_for_retry(&path, &e, &mut retry, "process")
-                            .await
-                        {
-                            return;
-                        }
+                }
+            }
+        }
+
+        loop {
+            let reader = match self.src_store.open_reader(&path).await {
+                Ok(reader) => reader,
+                Err(e) => {
+                    if retry_state.should_retry(&e, "open reader for", &path) {
+                        retry_state.backoff().await;
+                        continue;
                     }
+                    self.operation.record_failure(&path).await;
+                    return;
+                }
+            };
+
+            // Process the file - process_object cleans up partial files on error
+            match self.operation.process_object(&path, reader).await {
+                Ok(()) => {
+                    debug!("Processed: {}", path);
+                    self.operation.record_success(&path);
+                    return;
+                }
+                Err(e) => {
+                    if retry_state.should_retry(&e, "process", &path) {
+                        retry_state.backoff().await;
+                        continue;
+                    }
+                    self.operation.record_failure(&path).await;
+                    return;
                 }
             }
         }
