@@ -1,8 +1,9 @@
 use crate::history_format;
 /// Scan operation - validates that files exist
 use crate::pipeline::{async_trait, Operation};
-use crate::storage::{BoxedAsyncRead, Error as StorageError, StorageRef};
+use crate::storage::{Error as StorageError, StorageRef};
 use crate::utils::{compute_checkpoint_bounds, fetch_well_known_history_file, ArchiveStats};
+use opendal::{Buffer, Reader};
 use thiserror::Error;
 
 /// Scan operation errors
@@ -27,14 +28,25 @@ pub struct ScanOperation {
     // User-specified arguments from CLI
     low: Option<u32>,
     high: Option<u32>,
+
+    // Retry configuration for source fetches
+    max_retries: u32,
+    retry_min_delay_ms: u64,
 }
 
 impl ScanOperation {
-    pub async fn new(low: Option<u32>, high: Option<u32>) -> Result<Self, Error> {
+    pub async fn new(
+        low: Option<u32>,
+        high: Option<u32>,
+        max_retries: u32,
+        retry_min_delay_ms: u64,
+    ) -> Result<Self, Error> {
         Ok(Self {
             stats: ArchiveStats::new(),
             low,
             high,
+            max_retries,
+            retry_min_delay_ms,
         })
     }
 }
@@ -44,12 +56,11 @@ impl Operation for ScanOperation {
     async fn get_checkpoint_bounds(
         &self,
         source: &StorageRef,
-        max_retries: u32,
-        initial_backoff_ms: u64,
     ) -> Result<(u32, u32), crate::pipeline::Error> {
-        let source_state = fetch_well_known_history_file(source, max_retries, initial_backoff_ms)
-            .await
-            .map_err(|e| crate::pipeline::Error::ScanOperation(Error::Utils(e)))?;
+        let source_state =
+            fetch_well_known_history_file(source, self.max_retries, self.retry_min_delay_ms)
+                .await
+                .map_err(|e| crate::pipeline::Error::ScanOperation(Error::Utils(e)))?;
         let source_checkpoint =
             history_format::round_to_lower_checkpoint(source_state.current_ledger);
 
@@ -57,15 +68,16 @@ impl Operation for ScanOperation {
             .map_err(|e| crate::pipeline::Error::ScanOperation(Error::Utils(e)))
     }
 
-    async fn process_object(
-        &self,
-        _path: &str,
-        _reader: BoxedAsyncRead,
-    ) -> Result<(), StorageError> {
+    async fn process_object(&self, _path: &str, _reader: Reader) -> Result<(), StorageError> {
         // For now, scan just checks existence.
         unreachable!(
             "ScanOperation uses existence_check_only(), process_object should not be called"
         )
+    }
+
+    async fn process_buffer(&self, path: &str, _buffer: Buffer) {
+        // Scan doesn't write - just record that we successfully downloaded and parsed the history
+        self.stats.record_success(path);
     }
 
     fn record_success(&self, path: &str) {
@@ -74,10 +86,6 @@ impl Operation for ScanOperation {
 
     async fn record_failure(&self, path: &str) {
         self.stats.record_failure(path).await;
-    }
-
-    fn record_retry(&self) {
-        self.stats.record_retry();
     }
 
     fn record_skipped(&self, _path: &str) {
