@@ -1,5 +1,6 @@
 //! Tests for command-line interface validation and error handling
 
+use super::utils::file_url_from_path;
 use crate::test_helpers::{run_mirror as cmd_mirror_run, MirrorConfig};
 use rstest::rstest;
 
@@ -36,9 +37,7 @@ async fn test_mirror_rejects_malformed_url(
     let result = cmd_mirror_run(config).await;
     assert!(
         result.is_err(),
-        "Should reject malformed {} URL: '{}'",
-        position,
-        bad_url
+        "Should reject malformed {position} URL: '{bad_url}'"
     );
 }
 
@@ -59,8 +58,7 @@ async fn test_rejects_nonexistent_source(#[case] operation: &str) {
 
     assert!(
         result.is_err(),
-        "{} should reject nonexistent source",
-        operation
+        "{operation} should reject nonexistent source"
     );
 }
 
@@ -73,15 +71,14 @@ async fn test_mirror_creates_destination_if_not_exists() {
     let temp_base = TempDir::new().expect("Failed to create temp base");
     let dest_path = temp_base.path().join("new-dest-directory");
 
-    let src = format!("file://{}", test_archive_path().display());
-    let dst = format!("file://{}", dest_path.display());
+    let src = file_url_from_path(&test_archive_path());
+    let dst = file_url_from_path(&dest_path);
     let config = MirrorConfig::new(&src, &dst).skip_optional().high(63);
 
     let result = cmd_mirror_run(config).await;
     assert!(
         result.is_ok(),
-        "Should create destination directory if it doesn't exist, got error: {:?}",
-        result
+        "Should create destination directory if it doesn't exist, got error: {result:?}"
     );
 
     assert!(dest_path.exists(), "Destination directory should exist");
@@ -100,28 +97,23 @@ async fn test_rejects_low_greater_than_high(#[case] operation: &str) {
     use crate::test_helpers::{run_scan, ScanConfig};
     use tempfile::TempDir;
 
-    let src = format!("file://{}", test_archive_path().display());
+    let src = file_url_from_path(&test_archive_path());
 
     let result = match operation {
         "scan" => run_scan(ScanConfig::new(&src).low(2000).high(1000)).await,
         "mirror" => {
             let temp_dir = TempDir::new().unwrap();
-            let dst = format!("file://{}", temp_dir.path().join("dest").display());
+            let dst = file_url_from_path(&temp_dir.path().join("dest"));
             cmd_mirror_run(MirrorConfig::new(&src, &dst).low(2000).high(1000)).await
         }
         _ => unreachable!(),
     };
 
-    assert!(
-        result.is_err(),
-        "{} should reject when low > high",
-        operation
-    );
+    assert!(result.is_err(), "{operation} should reject when low > high");
 
     let err_msg = result.unwrap_err().to_string();
     assert!(
         err_msg.contains("low checkpoint") && err_msg.contains("is greater than high checkpoint"),
-        "Error should indicate low > high, got: {}",
-        err_msg
+        "Error should indicate low > high, got: {err_msg}"
     );
 }
